requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pandas==2.2.2
ta==0.11.0
python-binance==1.0.19
openpyxl==3.1.5